tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-flow-invalid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Grid Lanes: grid-lanes-flow parsing</title>
  <link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
  <script>
    test_invalid_value("grid-lanes-flow", "nowrap row nowrap");
    test_invalid_value("grid-lanes-flow", "column wrap column");
    test_invalid_value("grid-lanes-flow", "reverse-column column");
    test_invalid_value("grid-lanes-flow", "reverse-row, row");
    test_invalid_value("grid-lanes-flow", "normal row-reverse");
    test_invalid_value("grid-lanes-flow", "reverse column");
    test_invalid_value("grid-lanes-flow", "auto, auto");
  </script>
</body>
</html>
